+++ src/eez/gui/app_context.h
@@ -82,10 +82,10 @@ public:
     int getPreviousPageId();
     Page *getPreviousPage();
 
-    void pushSelectFromEnumPage(const data::EnumItem *enumDefinition, uint8_t currentValue,
-                                bool (*disabledCallback)(uint8_t value), void (*onSet)(uint8_t));
+    void pushSelectFromEnumPage(const data::EnumItem *enumDefinition, uint16_t currentValue,
+                                bool (*disabledCallback)(uint16_t value), void (*onSet)(uint16_t));
     void pushSelectFromEnumPage(void (*enumDefinitionFunc)(data::DataOperationEnum operation, data::Cursor &cursor, data::Value &value), 
-                                uint8_t currentValue, bool (*disabledCallback)(uint8_t value), void (*onSet)(uint8_t));
+                                uint16_t currentValue, bool (*disabledCallback)(uint16_t value), void (*onSet)(uint16_t));
 
     void replacePage(int pageId, Page *page = nullptr);
